Tidies up currencies
I wanted to know how to style the SVG with the text nicer. I found
this StackOverflow post https://stackoverflow.com/questions/37742099/how-do-you-align-text-with-svg-elements

I decided to create another fragment which would only render a single
svg and number. This would keep things cleaner and easier to work with.

After that was done, It enabled me to update the Currencies to only worry
about arranging things.