Add reduceAndRound function to getWealth
This function will reduce an array down to a number and round the
result to two decimal places.

It highlights a couple of new ideas, passing functions and generic types.

You'll notice the odd looking <T>. This is a generic type, its not too scary
it just means the caller gets to tell the function what the type is.

So reduceAndRound<Item> will set T to Item.

This gives you full access to types in your reduce function

It takes two arguments
  1. an array of T
         someArray: T[]

  2. A function with 2 arguments, a number and a T, and returns a number
         reduceFunc: (acc: number, item: T) => number)

We call reduce on the array and just dump in the function along with a 0, then do
the rounding you were already doing.

Hopefully this gives you some stuff to think about.

These are both powerful concepts.