Adds testing libraries (Read whole message)
I've added mocha & chai so we can start testing things.
There's some setup you'll need to do;

First install typescript and mocha globally
    npm install -g mocha typescript

Install the new dm-tools dependencies
    npm install

You should now be able to run the test with
    npm test

A quick look at whats happening when we run the tests.
Typescript doesn't run, so we need to convert it to JS
We run `tsc` as a pretest script. This compiles the TS into JS
and sticks it into a folder called js. This is whats run by mocha
and we .gitignore this dir so we don't commit it.