Adds a test driven version of armour class function
Part of the drive to split the character sheet generation into
seperate function is so we can test it. As it is now I'm scared to
change anything as I tested it all by checking it against dndbeyond.

As AC is a complex thing to  calculate I think testing here is mega
important. I've started off how I'd have started it to give you something
to follow.