Combines multiclass hit die if they are the same
If we had a multi class where both classes use the same hit dice then we
would want to display the combined total of that dice.

I've updated the hit dice calculation to combine the `max` and `used` of each
dice type before reforamtting the data to conform to the HitDice type.