Creates HpBar component
I'd like to reuse the HP bar for any pets/mounts so I've pulled out
the HP bar from Health and made it a component.

It takes two props `max` and `removed`. It is responsible for calculatig
the percentage, as well as setting the correct colour.

This broke some things in Health which were using the hpPercent to set
some styling options. I've replcaed that with an `isUnconscious` function
which I think better describes what we're saying.

I also moved the ThemeProvider in Health outside of the character loop.
It made sense that they would all share one theme.