Moves theme settings into Theme component
The styling of Health was better than other pages.
I've pulled the Heath Theme out and made it its own component
called Theme.

The entire app is then wrapped in this theme in Dashboard.tsx.
This ensures all child components get the same theming as each other
and gives a more consistent style to the app.