Adds creatures to Health page
I wanted to track the HP of player creatures, like Wilson.

It was simple enough to pull it from the dndbeyond data with
some simple mapping.

I did have to do some custom logic for wilson as his HP is based
on the artficer level and INT modifier. It's not too pretty and it'd
probably break for a multi-class, but its ok for now.

It also gave me a good oppertunity to reuse some components from the Health
refactor Ac and HpBar.