Adds health potions to Health page
I've taken your idea and ran with it. Adding healing potions
to the health page.

I installed the react-icons package which give us access to a
load of icons.

https://www.npmjs.com/package/react-icons

I've chosen to display an icon for each potion which a character is
carrying, I think they make make it easier to see from the other side
of the room.

I've also coloured them in the WoW colours to denote the strength
of the potion.

You will need to yarn install before you run this.