Moves creatures under avatar
I was worried about using too much vertical space and everything
not fitting on the screen.

I've moved the creatures to below the avatar to save some space.

I've also made a few tweaks to the avatar. I've set the width to always
be 100% of its container. I've also set the avatar section to always be
20% of the character. It makes it scale down a bit nicer.