Pulls out the getting of a file from getSummary
I want to play with creating graphs of thing like hp. In order
to do this we must store hp over time. I'm planning to create
a file which stores metrics for the party. This file will need
to be read and written each time we load the party.

I've split out the getting of a file from disk into getFile, which
given a path will get the file and return it as json. If the file
does not exist, it will be created as an empty json object.