Adds metrics
Adds the start of metrics tracking.

Every time we fetch data from dndbeyond we now store the XP & HP
of all characters in a json file. This json file is then loaded by
the metrics page to display graphs over time.

This is all still a bit rough and ready but it seems to basicaly
work.