Parse graph dates in handler
We need to turn the date strings into Date objects for the graph
to work. Previously I was doing this in the graphs themselves, this
is slow because we're doing it for each graph and on every rerender.

I've moved the logic into the handler so it happens once when we load
the data from disk.