Graphs handle deleted characters
As the number of datapoints for each character needs to equal the
number of entries in the xAxis deleted characters would break the
graphs.

The solution is to add `null` entries into deleted characters series

This is a good example of using a closure to add variables to a function
which we cannot change the signature of.

forEach will always call the supplied function for each element in the
array, but we also want to also give it numberOfDataPoints.

fillDataPoint returns a function which forEach can call with
numberOfDataPoints already set to what it is called with.