Adds logic to apply time periods to graphs
I wanted to be able to look at the last n hours of the metric data
ideally this should be selectable by the user. Turns out its hard
as the state is always lost during the character refresh.

I've moved most of the code for reducing the metrics into the handler.

The handler is called with a number of milliseconds which is converted
to a Date (now - millisecond window). We then iterate throough the x axis
timestampes and find the first instance of a time which is higher than our
Date, we then take the index of this Date.

We slice off everything before that index in all of the arrays. This
gives us our window of time.

If a character has been removed from the party they still would show up
due to us adding null for each metric.

Before the data is returned we remove any character which has all null
values in our time range. We do this by converting the array to a Set
(which only allows unique values) and then back to an array. Now we have
an array of unique values.

If it has a length of 1 and that value is null we remove tha character.

There's still quite a bit of work to do here, which may be solved by how
we refresh the data from dndbeyond.