Moves shared type definitions into their own file
It's bothered me for a bit that the DmToolsData types were defined in
the CharacterSheetProcessor with some in other places.

I've extracted them into dm-tools-data.types.ts Which gives us a single
place to define them.

I've also made ddb-data.types.ts which defines types we've made for the
dnd-beyond data.