LanguagesSelector works with an array of strings
I've updated the LanguagesSelector to work on an array of
strings. Now all we need to do is ad/remove strings from
the languages array when checkboxes are clicked.

This will reduce the number of hooks we need.

It does not yet update with the current languages and I've
run out of time.