Removes local hooks from LanguagesSelector
The reason this was getting hard was managing two sets of hooks which
were doing the same job.

Rather than making a new `languages` in LanguagesSelector it's much
cleaner to use the what's passed in. This gets everything working as
when the `languages` are updated this component rerenders showing the
correct state.