Replaces ReactDom render
I saw the following warning in the console

```
ReactDOM.render is no longer supported in React 18. Use createRoot instead.
Until you switch to the new API, your app will behave as if it’s running React 17.
```

It pointed to this link https://react.dev/blog/2022/03/08/react-18-upgrade-guide#updates-to-client-rendering-apis

I followed the process to migrate to the new method of loading
the app.