Fixes incorrect use of hook errors
The way we were loading data and refreshing it when it updated was
a bit janky and I think we were misusing useEffect.

I've updated it to a state where React seems happy and I think it makes
a little bit more sense.

On the initial page load we use useEffect to get the data we need then
set `pageRendered` to true so it won't do anything else.

I've moved the event handler out of the useEffect function and it just
works! I suspect this is what it was angry about.

This should remove the last of the errors from the app.