Reworks how characters are loaded by pages
By setting the event listener outside of useEffect it was being set
every time the component re-rendered. This is bad bacause it would
load data for every listener it had registered and register a new
listener each time.

I've removed the if statement in useEffect, by passing in an empty array the
function only runs once.

I've also been looking into how we process events. I noticed that after
you change page and refresh the data, all previous pages still update
this is due to the listeners still being registered.

useEffect can return a function which is used to clean up after itself
when the component is done. I wanted to look into how I could use this to
destroy the listener when we change page.

I've started to update preload.js to have a few functions which can be used
by anything. One of those is `receive` it will regiter a function to an event
and also return a function to destroy that listener. We can now use this in
useEffect to tidy up after ourselves.