Adds getParty function
Its kind of a faff to get all the charecters of the party.
We had to get the summary and then map over the IDs in an
async map. We were doing this all over the place.

I've replaced this with an async function of getParty()
now we can just call this and map over the DmToolsData[]