Uses custom hook and improves inventories
I added the custom hook to inventory as I had with the others.

I then took the time to improve the search function as I now understand
hooks better.

Search was always a bit janky, it would break if the data reloaded while
you were searching for something.

I've fixed that by storing the search term in state, so it persists between
reloads.

I've added a useEffect hook which will do the search if the search string changes
it will also re-run the search if the fullInventory changes.

This means that when the data is reloaded the search properly updates and searchs
the new data.