Logs everything in the terminal
Since I've been thinking a lot about logging it kind of bothered
me that we had two seperate streams of logs.

I've created some logger classes that we should use to log now.
In the backend we use Logger this will log to the terminal as usual
while formatting the log and adding a timestamp.

In the front end we use RenedererLogger. This class will log to
dev-tools as usual but will also send all log lines as events to the
backend to be logged to the terminal.

EventLogger is used to handle the log events sent from the front end and
will call the Logger class to log to the terminal.